Change name of superqueued_jobs to queued_jobs
The entire database is considered the super queue,
so calling a table superqueued_jobs is confusing.
A buffered job is a job that has been moved from
the buffer queue to the database and a queued job
is a job that has been processed by the superqueuer
and moved out of buffered_jobs.

Issue #2