Remove higher order implementation of generators
Ignorance made me believe that returning the generator
in a higher order function was required so the same
generator would not be used over and over, potentially
creating conflicts. In fact, it will not create a conflict
so using higher order functions to "fix" an issue that
is not there is in fact just complicating matters.

The higher order function has been removed to make things
more clear