Acknowledge RabbitMQ message even if job not found in DB
In some cases a job will be acknowledged in the DB but not
in RabbitMQ, which will cause the job to re-run when the
worker picks the message back up out of RabbitMQ.  When the
job is re-ran, the database queries to acknowledge the job
in the DB would fail because the job would no longer be in
buffered_jobs.  As a result the job again would not be ack'ed
in RabbitMQ.  This would cause the job to re-run over and over.

It would be best to acknowledge the message in RabbitMQ and
re-throw the exception so we know about the issue but also
allow for the issue to self-heal.

Issue #82