Move AMQP channel generation to Channel class
By having the AMQP channel generated in the Channel class
we can cleanly pass in the Connection to the Channel class.
Passing in the Connection to the Channel means that so long
as the Channel exists, the destructor of the Connection will
not be called and the Channel's AMQP Connection should be
maintained.