Implement in-memory message queue testing adapters
When we go to test the other components of the job queue,
we'd rather not build mocks but we also do not want to be
connecting to RabbitMQ for every test.  The in-memory MQ
adapters will allow us to test functionality that depends
on the MQ adapters using an in-memory solution that functions
like RabbitMQ does.