Generate adapter factory just in time in Testing\Config
The Config constructor requires an AdapterFactory and the
AdapterFactory constructor requires a Config. Up until now
the two testing classes have not yet needed to be used with
each other. We may want to stop supplying the adapter
factory through the config object, but for now let's use a
callable to generate the adapter factory so the Config
constructor does not directly require an AdapterFactory.