Correct semantics of DB transaction test
The test was testing to see if a transaction could
be used surrounding the buffering of a job, but the
purpose of the DB adapter's begin/commitTransaction()
is so that moving jobs from buffered_jobs to queued_jobs
can be batched in parallel with pushing the jobs back
to Rabbit.

This test covers a regression that was fixed in #183.