Update tests to use configured host for RabbitMQ
The RabbitMQ host was being hard-coded to localhost for some tests,
which does not work when RabbitMQ is hosted on another host
(READ: docker container). Using the hostname listed in the test config
for these tests will allow the tests to work just like the tests that
already work with a separate RabbitMQ host.